VariableComment: issue when handling props with union and intersection
This commit fixes an issue in the Squiz.Commenting.VariableComment sniff
that was causing it to wrongly report missing docblocks for properties
with union and intersection types.

This sniff has a list of tokens that it should skip when searching for
the end of a docblock comment that sits above a variable. The tokens for
the union and intersection types (T_TYPE_UNION and T_TYPE_INTERSECTION)
were missing in this list.

Adding them fixed the problem as now the code can properly identify the
end of the docblock comment when those types are present.