Tests/Tokenizer: tests to ensure scope is set correctly for T_SWITCH
This commit adds tests to safeguard against regressions related to
b24b96b which was added to fix one of the issues in
https://github.com/squizlabs/PHP_CodeSniffer/issues/497. b24b96b
originally added a test to a sniff test
(InlineControlStructureUnitTest.php) as back then there were no
Tokenizer tests. This test was added to ensure that
Tokenizer::recurseScopeMap() would correctly add scope information to
the T_SWITCH token when handling a `switch` that uses the alternative
syntax.

Now that the InlineControlStructure sniff doesn't listen for the
T_SWITCH token anymore, the original test added in b24b96b became
useless and thus was refactored in a previous commit to use different
control structure.