Tests/Tokenizer: tests for T_CASE and T_IF scope condition
This commit adds tests to safeguard against regressions related to
fddc61a which is part of
https://github.com/squizlabs/PHP_CodeSniffer/issues/497. fddc61a
originally added a test to a sniff test
(InlineControlStructureUnitTest.php) as back then there were no
Tokenizer tests. Those tests were added to ensure that
Tokenizer::recurseScopeMap() would correctly add scope information to
the `T_CASE` token when it shares the closer with `T_SWITCH` and to
`T_IF` when a nested `T_CASE` is missing shares the closer with
`T_SWITCH`.

Now that the InlineControlStructure sniff doesn't listen for the
T_SWITCH token anymore, the original test added in fddc61a became
useless and thus was refactored in a previous commit to use different
control structure.